Renamed the stack and made it a long long

+++ src/stack.h
@@ -3,10 +3,11 @@
 
 void push(const long data);
 void show_stack();
-long read_stack(size_t offset);
-long pop(size_t offset);
+long long read_stack(size_t offset);
+long long pop(size_t offset);
 size_t get_index();
 void clear_stack();
 void* push_string(char* str);
 void pop_string(void *str);
+
 #endif
